+++ source/lambda/utilsLayer/lib/exports.ts
@@ -265,3 +265,7 @@ export function getNotificationMutingStatus(
     }
   }
 }
+
+export async function sleep(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}