+++ source/resources/__tests__/hub.spec.ts
@@ -6,7 +6,24 @@ import { QuotaMonitorHub } from "../lib/hub.stack";
 import { App } from "aws-cdk-lib";
 
 describe("==Hub Stack Tests==", () => {
-  const app = new App();
+  const app = new App({
+    context: {
+      "SOLUTION_VERSION": "test_version",
+      "SOLUTION_NAME": "test_name",
+      "SOLUTION_ID": "SO0005",
+      "SOLUTION_BUCKET": "test_bucket",
+      "SOLUTION_TEMPLATE_BUCKET": "test_bucket",
+      "CUSTOM_USER_AGENT": "AwsSolution/SO0005/test_version",
+      "SEND_METRICS": "Yes",
+      "METRICS_ENDPOINT": "https://metrics.awssolutionsbuilder.com/generic",
+      "LOG_LEVEL": "info",
+      "APPLICATION_TYPE": "AWS-Solutions",
+      "APP_REG_HUB_NO_OU_APPLICATION_NAME": "QM_Hub",
+      "APP_REG_HUB_APPLICATION_NAME": "QM_Hub_Org",
+      "APP_REG_TA_SPOKE_APPLICATION_NAME": "QM_TA",
+      "APP_REG_SQ_SPOKE_APPLICATION_NAME": "QM_SQ"
+    },
+  });
   const stack = new QuotaMonitorHub(app, "QMHubStack");
   const template = Template.fromStack(stack);
 
@@ -99,23 +116,8 @@ describe("==Hub Stack Tests==", () => {
 
     it("should have an SNS Topic for SNSNotifier", () => {
       template.resourceCountIs("AWS::SNS::Topic", 1);
-      template.hasResource("AWS::SNS::Topic", {
-        Properties: {
-          KmsMasterKeyId: {
-            "Fn::Join": [
-              "",
-              [
-                "arn:",
-                { Ref: "AWS::Partition" },
-                ":kms:",
-                { Ref: "AWS::Region" },
-                ":",
-                { Ref: "AWS::AccountId" },
-                ":alias/aws/sns",
-              ],
-            ],
-          },
-        },
+      template.hasResourceProperties("AWS::SNS::Topic", {
+        KmsMasterKeyId: Match.anyValue(),
       });
     });
 
@@ -196,6 +198,36 @@ describe("==Hub Stack Tests==", () => {
         Type: "String",
         AllowedValues: [ "Organizations", "Hybrid" ],
       });
+    });
+
+    it("should have Service Catalog AppRegistry Application, ", () => {
+      template.resourceCountIs("AWS::ServiceCatalogAppRegistry::Application", 1);
+    });
+
+    it("should have Service Catalog AppRegistry AttributeGroup, ", () => {
+      template.resourceCountIs("AWS::ServiceCatalogAppRegistry::AttributeGroup", 1);
+    });
+
+    it("should have Service Catalog AppRegistry Resource Association, ", () => {
+      template.resourceCountIs("AWS::ServiceCatalogAppRegistry::ResourceAssociation", 1);
+      template.hasResource("AWS::ServiceCatalogAppRegistry::ResourceAssociation", {
+        Properties: {
+          Application: {
+            "Fn::GetAtt": [
+              "HubAppRegistryApplication3E8980C3",
+              "Id"
+            ]
+          },
+          Resource: {
+            "Ref": "AWS::StackId"
+          },
+          ResourceType: "CFN_STACK",
+        }
+      });
+    });
+
+    it("should have Service Catalog AppRegistry AttributeGroup Association, ", () => {
+      template.resourceCountIs("AWS::ServiceCatalogAppRegistry::AttributeGroupAssociation", 1);
     });
 
   });