+++ source/resources/__tests__/sq-spoke.spec.ts
@@ -6,7 +6,24 @@ import { QuotaMonitorSQSpoke } from "../lib/sq-spoke.stack";
 import { App } from "aws-cdk-lib";
 
 describe("==SQ-Spoke Stack Tests==", () => {
-  const app = new App();
+  const app = new App({
+    context: {
+      "SOLUTION_VERSION": "test_version",
+      "SOLUTION_NAME": "test_name",
+      "SOLUTION_ID": "SO0005",
+      "SOLUTION_BUCKET": "test_bucket",
+      "SOLUTION_TEMPLATE_BUCKET": "test_bucket",
+      "CUSTOM_USER_AGENT": "AwsSolution/SO0005/test_version",
+      "SEND_METRICS": "Yes",
+      "METRICS_ENDPOINT": "https://metrics.awssolutionsbuilder.com/generic",
+      "LOG_LEVEL": "info",
+      "APPLICATION_TYPE": "AWS-Solutions",
+      "APP_REG_HUB_NO_OU_APPLICATION_NAME": "QM_Hub",
+      "APP_REG_HUB_APPLICATION_NAME": "QM_Hub_Org",
+      "APP_REG_TA_SPOKE_APPLICATION_NAME": "QM_TA",
+      "APP_REG_SQ_SPOKE_APPLICATION_NAME": "QM_SQ"
+    },
+  });
   const stack = new QuotaMonitorSQSpoke(app, "SQSpokeStack", {});
   const template = Template.fromStack(stack);
 
@@ -109,6 +126,36 @@ describe("==SQ-Spoke Stack Tests==", () => {
       const allParams = template.findParameters("*", {});
       expect(allParams).toHaveProperty("NotificationThreshold");
       expect(allParams).toHaveProperty("MonitoringFrequency");
+    });
+
+    it("should have Service Catalog AppRegistry Application, ", () => {
+      template.resourceCountIs("AWS::ServiceCatalogAppRegistry::Application", 1);
+    });
+
+    it("should have Service Catalog AppRegistry AttributeGroup, ", () => {
+      template.resourceCountIs("AWS::ServiceCatalogAppRegistry::AttributeGroup", 1);
+    });
+
+    it("should have Service Catalog AppRegistry Resource Association, ", () => {
+      template.resourceCountIs("AWS::ServiceCatalogAppRegistry::ResourceAssociation", 1);
+      template.hasResource("AWS::ServiceCatalogAppRegistry::ResourceAssociation", {
+        Properties: {
+          Application: {
+            "Fn::GetAtt": [
+              "SQSpokeAppRegistryApplicationB3787B2B",
+              "Id"
+            ]
+          },
+          Resource: {
+            "Ref": "AWS::StackId"
+          },
+          ResourceType: "CFN_STACK",
+        }
+      });
+    });
+
+    it("should have Service Catalog AppRegistry AttributeGroup Association, ", () => {
+      template.resourceCountIs("AWS::ServiceCatalogAppRegistry::AttributeGroupAssociation", 1);
     });
   });
 });